+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add configuration `EndpointConfig::errorsAreClientSafe()` to propagate client-safety of endpoint errors
+
+### Changed
+
+- Move `Spawnia\Sailor\ResultErrorsException` to `Spawnia\Sailor\Error\ResultErrorsException`
+
+### Removed
+
+- Remove `Spawnia\Sailor\Response::assertErrorFree()`
+
 ## v0.15.0
 
 ### Added